table drops and deletes are now logged
